+++ src/store/modules/index.js
@@ -1,10 +1,12 @@
 import Assets from '@/store/modules/assets'
+import Collections from '@/store/modules/collections'
 import Page from '@/store/modules/page'
 import Pages from '@/store/modules/pages'
 import User from '@/store/modules/user'
 
 export default {
   Assets,
+  Collections,
   Page,
   Pages,
   User,